pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.33.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
